test modification for array with no element
